arrumado alguns redirecionametnos

+++ curso-jsp/src/main/java/dao/DAOUsuarioRepository.java
@@ -19,7 +19,7 @@ public class DAOUsuarioRepository {
 
 	public ModelLogin gravarUsuario(ModelLogin objeto) throws Exception {
 
-		if (objeto.isNovo()) {/*Grava um novo*/
+		if (objeto.isNovo()) {/* Grava um novo */
 
 			String sql = "INSERT INTO model_login(login, senha, nome, email) VALUES (?, ?, ?, ?);";
 			PreparedStatement statement = connection.prepareStatement(sql);
@@ -31,72 +31,69 @@ public class DAOUsuarioRepository {
 
 			statement.execute();
 			connection.commit();
-		}else {
-			String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=? WHERE id="+objeto.getId()+";";
+		} else {
+			String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=? WHERE id=" + objeto.getId() + ";";
 			PreparedStatement statement = connection.prepareStatement(sql);
-			
+
 			statement.setString(1, objeto.getLogin());
 			statement.setString(2, objeto.getSenha());
 			statement.setString(3, objeto.getNome());
 			statement.setString(4, objeto.getEmail());
-			
+
 			statement.executeUpdate();
 			connection.commit();
 		}
 		return this.consultaUsuario(objeto.getLogin());
 
 	}
-	
-	public List<ModelLogin> consultaUsuarioList (String nome) throws Exception{
-				
+
+	public List<ModelLogin> consultaUsuarioList() throws Exception {
+
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
-		
-				
-		String sql = "select * from model_login where upper(nome) like upper(?)";
+
+		String sql = "select * from model_login where useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
-		statement.setString(1,"%" + nome + "%");
-		
+
 		ResultSet resultado = statement.executeQuery();
-		
-		while(resultado.next()) {
+
+		while (resultado.next()) { /* percorrer as linhas de resultado do SQL */
+
 			ModelLogin modelLogin = new ModelLogin();
-			
+
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setId(resultado.getLong("id"));
 			modelLogin.setLogin(resultado.getString("login"));
 			modelLogin.setNome(resultado.getString("nome"));
-			
-			retorno.add(modelLogin);
+			// modelLogin.setSenha(resultado.getString("senha"));
 
+			retorno.add(modelLogin);
 		}
-		
-		
+
 		return retorno;
 	}
-	
-	public List<ModelLogin> consultaUsuarioList() throws Exception{
-		
+
+	public List<ModelLogin> consultaUsuarioList(String nome) throws Exception {
+
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
-		
-				
-		String sql = "select * from model_login";
+
+		String sql = "select * from model_login where upper(nome) like upper(?) and useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
-				
+		statement.setString(1, "%" + nome + "%");
+
 		ResultSet resultado = statement.executeQuery();
-		
-		while(resultado.next()) {
+
+		while (resultado.next()) {
 			ModelLogin modelLogin = new ModelLogin();
-			
+
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setId(resultado.getLong("id"));
 			modelLogin.setLogin(resultado.getString("login"));
 			modelLogin.setNome(resultado.getString("nome"));
-			
+
 			retorno.add(modelLogin);
 
 		}
-		
-		
+
 		return retorno;
 	}
 
@@ -104,7 +101,7 @@ public class DAOUsuarioRepository {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where upper(login) = upper('" + login + "')";
+		String sql = "select * from model_login where upper(login) = upper('" + login + "') and useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 
 		ResultSet resultado = statement.executeQuery();
@@ -120,12 +117,12 @@ public class DAOUsuarioRepository {
 		return modelLogin;
 
 	}
-	
+
 	public ModelLogin consultaUsuarioID(String id) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where id = ? ";
+		String sql = "select * from model_login where id = ? and useradmin is false";
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setLong(1, Long.parseLong(id));
 
@@ -156,17 +153,16 @@ public class DAOUsuarioRepository {
 		return resultado.getBoolean("existe");
 
 	}
-	
-	public void deletarUser(String idUser) throws Exception{
-		String sql = "DELETE FROM model_login WHERE id =?;";
-		
+
+	public void deletarUser(String idUser) throws Exception {
+		String sql = "DELETE FROM model_login WHERE id =? and useradmin is false;";
+
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setLong(1, Long.parseLong(idUser));
-		
+
 		statement.executeUpdate();
 		connection.commit();
 
-				
 	}
 
 }
